backend/requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-GraphQL==2.0.1
graphene-sqlalchemy==2.2.2